Reorganize the packages and avoid the confusion between module and function with same name
