--- src/im2im/time_cost/__init__.py
@@ -1,2 +1,0 @@
-from .test_image_util import random_test_image_and_expected, is_image_equal
-from .time_cost_measure import time_cost, time_cost_in_kg
